requirements.txt
#langchain
#openai
#scholarly
#spacy

datasets==2.19.1
langchain-pinecone>=0.2.0
langchain-openai>=0.1.9
langchain>=0.3.11
langchain-core>=0.3.24
langgraph>=0.1.1
semantic-router==0.0.48
serpapi==0.1.5
google-search-results==2.4.2
langchain-groq>=0.2.1
langchain-community>=0.3.11
graphviz
pygraphviz
# pkgconfig
sentence-transformers